Removed compiler warnings
git-svn-id: https://icmake.svn.sourceforge.net/svnroot/icmake/trunk@1179 dcc949cb-2a1c-0410-b67a-e7e5ca830264

+++ icmake/comp/stackfra.c
@@ -11,7 +11,7 @@ ESTRUC_ *stackframe(E_TYPE_ type)
 {
     frame.type = type;
 
-    switch (type)
+    switch ((int)type)
     {
         case e_int | e_const:
             frame.evalue = atoi(lexstring);
